+++ good2travel/src/main/java/com/spot/good2travel/repository/FcmRepository.java
@@ -1,5 +1,6 @@
-package com.spot.good2travel.common.fcm;
+package com.spot.good2travel.repository;
 
+import com.spot.good2travel.domain.Fcm;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
